Code optimization

+++ Twitter-OAuth/TwitterOAuthViewController/oauthconsumer/Crypto/hmac.c
@@ -33,7 +33,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-void hmac_sha1(const u_int8_t *inText, size_t inTextLength, u_int8_t* inKey, size_t inKeyLength, u_int8_t *outDigest)
+void hmac_sha1(const u_int8_t *inText, int32_t inTextLength, u_int8_t* inKey, int32_t inKeyLength, u_int8_t *outDigest)
 {
 #define B 64
 #define L 20
